+++ Events/JumpedEvent.cs
@@ -32,6 +32,8 @@ namespace EddiEvents
             VARIABLES.Add("population", "The population of the system to which the commander has jumped");
             VARIABLES.Add("destination", "The route destination system, if any");
             VARIABLES.Add("destdistance", "The distance to the destination system)");
+            VARIABLES.Add("factions", "A list of faction objects describing the factions in the star system");
+            VARIABLES.Add("conflicts", "A list of conflict objects describing any conflicts between factions in the star system");
         }
 
         public string system { get; private set; }
@@ -67,7 +69,9 @@ namespace EddiEvents
 
         public List<Faction> factions { get; private set; }
 
-        // Faction properties
+        public List<Conflict> conflicts { get; private set; }
+
+        // Controlling faction properties
         public string faction => controllingfaction?.name;
         public string factionstate => (controllingfaction?.presences.FirstOrDefault(p => p.systemName == system)?.FactionState ?? FactionState.None).localizedName;
         public string allegiance => (controllingfaction?.Allegiance ?? Superpower.None).localizedName;
@@ -81,7 +85,7 @@ namespace EddiEvents
         public SecurityLevel securityLevel { get; private set; } = SecurityLevel.None;
         public FactionState factionState { get; private set; } = FactionState.None;
 
-        public JumpedEvent(DateTime timestamp, string system, long systemAddress, decimal x, decimal y, decimal z, string star, decimal distance, decimal fuelused, decimal fuelremaining, int? boostUsed, Faction controllingfaction, List<Faction> factions, Economy economy, Economy economy2, SecurityLevel security, long? population, string destination, decimal destdistance) : base(timestamp, NAME)
+        public JumpedEvent(DateTime timestamp, string system, long systemAddress, decimal x, decimal y, decimal z, string star, decimal distance, decimal fuelused, decimal fuelremaining, int? boostUsed, Faction controllingfaction, List<Faction> factions, List<Conflict> conflicts, Economy economy, Economy economy2, SecurityLevel security, long? population, string destination, decimal destdistance) : base(timestamp, NAME)
         {
             this.system = system;
             this.systemAddress = systemAddress;
@@ -95,6 +99,7 @@ namespace EddiEvents
             this.boostused = boostUsed;
             this.controllingfaction = controllingfaction;
             this.factions = factions;
+            this.conflicts = conflicts;
             this.Economy = (economy ?? Economy.None);
             this.Economy2 = (economy2 ?? Economy.None);
             this.securityLevel = (security ?? SecurityLevel.None);